4th Bug fixed: Incorrect function call
I have fixed the incorrect function call. In the JS code the (jsConcepts) parameter was called twice and that did not make sense to me. The findIntersection function is used if you want to find the common elements between parameters so using the same function call will only bring out the same results.